requirements.txt
typer
networkx
PyYAML
pytest